refactor(generate): (re)moved files to keep it DRY
