+++ src/generate/format-contributor.js
@@ -5,7 +5,7 @@ const avatarTemplate = _.template(
   '<img src="<%= contributor.avatar_url %>" width="<%= options.imageSize %>px;" alt="<%= name %>"/>',
 )
 const avatarBlockTemplate = _.template(
-  '[<%= avatar %><br /><sub><b><%= name %></b></sub>](<%= contributor.profile %>)',
+  '<a href="<%= contributor.profile %>"><%= avatar %><br /><sub><b><%= name %></b></sub></a>',
 )
 const avatarBlockTemplateNoProfile = _.template(
   '<%= avatar %><br /><sub><b><%= name %></b></sub>',
